requirements.txt
# Core dependencies for T-CRIS
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Survival analysis
lifelines>=0.27.8
scikit-survival>=0.22.0

# Machine Learning
scikit-learn>=1.3.0

# Deep Learning
torch>=2.0.0
pytorch-lightning>=2.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Framework
streamlit>=1.28.0

# Interpretability
shap>=0.43.0

# Utilities
loguru>=0.7.0
tqdm>=4.66.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# LLM Integration
groq>=0.33.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
